refactor : Redis rate limiter 공통 카운터 처리 추출
