+++ src/main/java/com/jinju/jinjuwiki/domain/auth/service/RedisEmailVerificationVerifyAttemptLimiter.java
@@ -2,12 +2,10 @@ package com.jinju.jinjuwiki.domain.auth.service;
 
 import com.jinju.jinjuwiki.global.error.BusinessException;
 import com.jinju.jinjuwiki.global.error.ErrorCode;
+import com.jinju.jinjuwiki.global.config.RedisCounterSupport;
 import java.time.Duration;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.data.redis.core.StringRedisTemplate;
-import org.springframework.data.redis.core.script.DefaultRedisScript;
 import org.springframework.stereotype.Component;
 
 // 이메일 인증코드 검증 실패 제한 Redis 클래스
@@ -16,9 +14,8 @@ import org.springframework.stereotype.Component;
 public class RedisEmailVerificationVerifyAttemptLimiter {
 
     private static final String VERIFY_FAILURE_KEY_PREFIX = "auth:email-verify-failure";
-    private static final DefaultRedisScript<Long> INCREMENT_WITH_TTL_SCRIPT = createIncrementWithTtlScript();
 
-    private final StringRedisTemplate stringRedisTemplate;
+    private final RedisCounterSupport redisCounterSupport;
 
     @Value("${app.auth.email-verification-verify-attempt-limit.count:5}")
     private long verifyFailureLimitCount;
@@ -28,7 +25,7 @@ public class RedisEmailVerificationVerifyAttemptLimiter {
 
     // 이메일 인증코드 검증 가능 여부 확인 메서드
     public void validateAllowed(String email) {
-        String failureCount = stringRedisTemplate.opsForValue().get(createRedisKey(email));
+        String failureCount = redisCounterSupport.get(createRedisKey(email));
         if (failureCount == null) {
             return;
         }
@@ -40,11 +37,7 @@ public class RedisEmailVerificationVerifyAttemptLimiter {
 
     // 이메일 인증코드 검증 실패 누적 메서드
     public long recordFailure(String email) {
-        Long currentCount = stringRedisTemplate.execute(
-                INCREMENT_WITH_TTL_SCRIPT,
-                List.of(createRedisKey(email)),
-                String.valueOf(getVerifyFailureWindow().toSeconds())
-        );
+        Long currentCount = redisCounterSupport.incrementWithTtl(createRedisKey(email), getVerifyFailureWindow());
 
         if (currentCount == null) {
             throw new BusinessException(ErrorCode.INVALID_INPUT);
@@ -55,7 +48,7 @@ public class RedisEmailVerificationVerifyAttemptLimiter {
 
     // 이메일 인증코드 검증 실패 초기화 메서드
     public void reset(String email) {
-        stringRedisTemplate.delete(createRedisKey(email));
+        redisCounterSupport.delete(createRedisKey(email));
     }
 
     // 검증 실패 Redis 키 생성 메서드
@@ -66,19 +59,5 @@ public class RedisEmailVerificationVerifyAttemptLimiter {
     // 이메일 인증코드 검증 실패 제한 시간 계산 메서드
     private Duration getVerifyFailureWindow() {
         return Duration.ofSeconds(verifyFailureWindowSeconds);
-    }
-
-    // TTL 포함 증가 Lua 스크립트 생성 메서드
-    private static DefaultRedisScript<Long> createIncrementWithTtlScript() {
-        DefaultRedisScript<Long> script = new DefaultRedisScript<>();
-        script.setResultType(Long.class);
-        script.setScriptText("""
-                local current = redis.call('INCR', KEYS[1])
-                if current == 1 then
-                    redis.call('EXPIRE', KEYS[1], ARGV[1])
-                end
-                return current
-                """);
-        return script;
     }
 }
