+++ src/main/java/com/jinju/jinjuwiki/domain/auth/service/RedisPasswordResetRequestRateLimiter.java
@@ -2,12 +2,10 @@ package com.jinju.jinjuwiki.domain.auth.service;
 
 import com.jinju.jinjuwiki.global.error.BusinessException;
 import com.jinju.jinjuwiki.global.error.ErrorCode;
+import com.jinju.jinjuwiki.global.config.RedisCounterSupport;
 import java.time.Duration;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.data.redis.core.StringRedisTemplate;
-import org.springframework.data.redis.core.script.DefaultRedisScript;
 import org.springframework.stereotype.Component;
 
 // 비밀번호 재설정 요청 제한 Redis 클래스
@@ -16,9 +14,8 @@ import org.springframework.stereotype.Component;
 public class RedisPasswordResetRequestRateLimiter {
 
     private static final String RESET_LIMIT_KEY_PREFIX = "auth:password-reset";
-    private static final DefaultRedisScript<Long> INCREMENT_WITH_TTL_SCRIPT = createIncrementWithTtlScript();
 
-    private final StringRedisTemplate stringRedisTemplate;
+    private final RedisCounterSupport redisCounterSupport;
 
     @Value("${app.auth.password-reset-request-rate-limit.count:5}")
     private long resetLimitCount;
@@ -28,11 +25,7 @@ public class RedisPasswordResetRequestRateLimiter {
 
     // 비밀번호 재설정 요청 제한 확인 메서드
     public void validateAllowed(String email) {
-        Long currentCount = stringRedisTemplate.execute(
-                INCREMENT_WITH_TTL_SCRIPT,
-                List.of(createRedisKey(email)),
-                String.valueOf(getResetLimitWindow().toSeconds())
-        );
+        Long currentCount = redisCounterSupport.incrementWithTtl(createRedisKey(email), getResetLimitWindow());
 
         if (currentCount == null || currentCount > resetLimitCount) {
             throw new BusinessException(ErrorCode.INVALID_INPUT);
@@ -47,19 +40,5 @@ public class RedisPasswordResetRequestRateLimiter {
     // 비밀번호 재설정 요청 제한 시간 계산 메서드
     private Duration getResetLimitWindow() {
         return Duration.ofSeconds(resetLimitWindowSeconds);
-    }
-
-    // TTL 포함 증가 Lua 스크립트 생성 메서드
-    private static DefaultRedisScript<Long> createIncrementWithTtlScript() {
-        DefaultRedisScript<Long> script = new DefaultRedisScript<>();
-        script.setResultType(Long.class);
-        script.setScriptText("""
-                local current = redis.call('INCR', KEYS[1])
-                if current == 1 then
-                    redis.call('EXPIRE', KEYS[1], ARGV[1])
-                end
-                return current
-                """);
-        return script;
     }
 }
