+++ src/main/java/com/jinju/jinjuwiki/domain/auth/service/RedisPasswordResetVerifyAttemptLimiter.java
@@ -1,11 +1,9 @@
 package com.jinju.jinjuwiki.domain.auth.service;
 
+import com.jinju.jinjuwiki.global.config.RedisCounterSupport;
 import java.time.Duration;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.data.redis.core.StringRedisTemplate;
-import org.springframework.data.redis.core.script.DefaultRedisScript;
 import org.springframework.stereotype.Component;
 
 // 비밀번호 재설정 인증코드 확인 실패 제한 Redis 클래스
@@ -14,9 +12,8 @@ import org.springframework.stereotype.Component;
 public class RedisPasswordResetVerifyAttemptLimiter {
 
     private static final String PASSWORD_RESET_VERIFY_FAILURE_KEY_PREFIX = "auth:password-reset-verify-failure";
-    private static final DefaultRedisScript<Long> INCREMENT_WITH_TTL_SCRIPT = createIncrementWithTtlScript();
 
-    private final StringRedisTemplate stringRedisTemplate;
+    private final RedisCounterSupport redisCounterSupport;
 
     @Value("${app.auth.password-reset-verify-attempt-limit.count:5}")
     private long verifyFailureLimitCount;
@@ -26,7 +23,7 @@ public class RedisPasswordResetVerifyAttemptLimiter {
 
     // 비밀번호 재설정 인증코드 확인 가능 여부 메서드
     public boolean isAllowed(String email) {
-        String failureCount = stringRedisTemplate.opsForValue().get(createRedisKey(email));
+        String failureCount = redisCounterSupport.get(createRedisKey(email));
         if (failureCount == null) {
             return true;
         }
@@ -36,11 +33,7 @@ public class RedisPasswordResetVerifyAttemptLimiter {
 
     // 비밀번호 재설정 인증코드 확인 실패 누적 메서드
     public long recordFailure(String email) {
-        Long currentCount = stringRedisTemplate.execute(
-                INCREMENT_WITH_TTL_SCRIPT,
-                List.of(createRedisKey(email)),
-                String.valueOf(getVerifyFailureWindow().toSeconds())
-        );
+        Long currentCount = redisCounterSupport.incrementWithTtl(createRedisKey(email), getVerifyFailureWindow());
 
         if (currentCount == null) {
             throw new IllegalStateException("비밀번호 재설정 인증코드 실패 횟수 누적에 실패했습니다.");
@@ -51,7 +44,7 @@ public class RedisPasswordResetVerifyAttemptLimiter {
 
     // 비밀번호 재설정 인증코드 확인 실패 초기화 메서드
     public void reset(String email) {
-        stringRedisTemplate.delete(createRedisKey(email));
+        redisCounterSupport.delete(createRedisKey(email));
     }
 
     private String createRedisKey(String email) {
@@ -65,18 +58,5 @@ public class RedisPasswordResetVerifyAttemptLimiter {
     // 비밀번호 재설정 인증코드 확인 실패 제한 도달 여부 확인 메서드
     public boolean hasReachedLimit(long failureCount) {
         return failureCount >= verifyFailureLimitCount;
-    }
-
-    private static DefaultRedisScript<Long> createIncrementWithTtlScript() {
-        DefaultRedisScript<Long> script = new DefaultRedisScript<>();
-        script.setResultType(Long.class);
-        script.setScriptText("""
-                local current = redis.call('INCR', KEYS[1])
-                if current == 1 then
-                    redis.call('EXPIRE', KEYS[1], ARGV[1])
-                end
-                return current
-                """);
-        return script;
     }
 }
