+++ src/main/java/com/jinju/jinjuwiki/global/config/RedisCounterSupport.java
@@ -1,0 +1,47 @@
+package com.jinju.jinjuwiki.global.config;
+
+import java.time.Duration;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.core.StringRedisTemplate;
+import org.springframework.data.redis.core.script.DefaultRedisScript;
+import org.springframework.stereotype.Component;
+
+// Redis 카운터 증감/TTL 공통 처리 컴포넌트
+@Component
+@RequiredArgsConstructor
+public class RedisCounterSupport {
+
+    private static final DefaultRedisScript<Long> INCREMENT_WITH_TTL_SCRIPT = createIncrementWithTtlScript();
+
+    private final StringRedisTemplate stringRedisTemplate;
+
+    public Long incrementWithTtl(String key, Duration ttl) {
+        return stringRedisTemplate.execute(
+                INCREMENT_WITH_TTL_SCRIPT,
+                List.of(key),
+                String.valueOf(ttl.toSeconds())
+        );
+    }
+
+    public String get(String key) {
+        return stringRedisTemplate.opsForValue().get(key);
+    }
+
+    public void delete(String key) {
+        stringRedisTemplate.delete(key);
+    }
+
+    private static DefaultRedisScript<Long> createIncrementWithTtlScript() {
+        DefaultRedisScript<Long> script = new DefaultRedisScript<>();
+        script.setResultType(Long.class);
+        script.setScriptText("""
+                local current = redis.call('INCR', KEYS[1])
+                if current == 1 then
+                    redis.call('EXPIRE', KEYS[1], ARGV[1])
+                end
+                return current
+                """);
+        return script;
+    }
+}
